Make EarlGrey and Functional tests work for deployment target <= 10.0.
Remove use of APIS deprecated on iOS 10 deployment target:
Replace use of OSAtomic* APIs with methods from stdatomic.h.
Use notifyWhenInteractionChangesUsingBlock as a replacement for
notifyWhenInteractionEndsUsingBlock on iOS 10.0 and higher.

Disable tests that fail when deployment target is 9.0 or higher:
Disable failing tests that don't work when deployment target is 9.0 or
higher.
testSynchronizationWorksWithNSURLConnection relies upon a deprecated API,
so this test doesn't need to be run on iOS 9.0 or higher.
testStyledAlertView induces an alert on completion, which causes other tests
to fail, if the deployment target is 9.0 or higher.