Support PDF Display in UIWebViews
In a webview with HTML content, there is a valid view at _webView._internal.browserView, but the view is always empty. This causes EG's UIWebView Idling Resource to assume the webview is waiting to render.

In a webview with a PDF, there is a view at _webView._internal.pdfHandler.pdfView. This view *only* exists after the PDF is rendered (AFAICT) and does not exist if the webview is rendering something other than a PDF.

To fix, check if there is a rendered PDF in the webview before performing HTML rendering checks.