Fix GREYTestApplicationDistantObject's hostPort/hostBackground getter.
The old way doesn't work because KVO will force the getter of a property to be called before executing the setter. As the result, we can't set up KVO expectation inside the getter to wait for setter to happen.

To resolve this issue, we will abandon the KVO expectation, and use our own runloop spinner to wait for the setter instead.

PiperOrigin-RevId: 301929561