Adds C helper functions to transfer array's hosting process in EarlGrey.
We noticed Apple lacks of NSArray constructor that is compatible with eDO when users want to transit the owned process of an array from local process to a remote process, or the reverse direction. As a workaround, users create their own helper.

The EG supported helper functions will be used by customers who not yet create their own helper for this purpose, and have their arrays being used by the correct process. Thus they will not encounter the `isKindOfClass:` process mismatch issue.

PiperOrigin-RevId: 332481771