Fix isKindOfClass check in Diagnostics.
The grey_allOf and grey_anyOf matchers can be defined in test process, in which case `[matcher isKindOfClass:[foo class]]` in the app-side will fail. To fix this, we use the `isSubclassOfClass:` to avoid proxy.

Also remove the `isEqual:` check on arrays from different processes, as it also triggers `isKindOfClass:` on the parameter.

PiperOrigin-RevId: 333534321