Expose private API and invoke on it directly.
The function pointer returned from methodForSelector is not reliable way to invoke: the return value doesn't align with the call frame. It may be due to how msgSend is refactored in the latest runtime. Switching to a traditional way to invoke a private method by exposing it in the category interface.

PiperOrigin-RevId: 359705564