Remove the edge pan gesture safe bound from scrolling area.
Swiping from the edge can cause the pan gesture which usually triggers navigation behaviors. The scrolling action placed touch points outside this area to avoid the confusion, but this already caused the incorrect calculation due to private API usage, as well as this doesn't seem to be needed as the touch points are already smoothed out the other possible gestures and should only trigger the scrolling.

PiperOrigin-RevId: 360786170