Replace fishhook with interposer for simulator tests after iOS 12.
Fishhook becomes flaky and widely affects presubmit tests. Interposer is an existing approach that was previously applied for sanitizer tests only. We extend its usage to all simulator tests after iOS 12 and TGP works fine.

Interposer also works for the real device tests and can fix an iOS 15 issue. However, we don't turn it on right now because it will affect the start-up latency of the performance test.

This is a breaking change to the UI tests that use pre-built test rig, thus we add a test-env `GREY_LEGACY_DISPATCH_QUEUE_TRACKER` to allow pre-built test rig continue using fishhook.

PiperOrigin-RevId: 380921521