Add tolerance of one more touch-point for scroll failure detection.
From users' UI test, we found 3 touch move is not enough to trigger the initial change of the content offset of a scroll view. There are cases that the gesture is recognized as scroll at the 3rd touch-move without generating any offset. Adding one more touch-move will generate the offset and avoid this case to be identified as a scroll failure.

PiperOrigin-RevId: 455708239