Track screen rotation using UIViewController events.
The track is done on the deprecated UIViewController -willRotateToInterfaceOrientation:duration: and -didRotateFromInterfaceOrientation:. As of iOS 16, both methods are still actively used by `_UIWindowRotationAnimationController`, which is the controller for the app's rotation animation.

Existing tests of the rotation API can be used to monitor the tracked APIs to make sure they are still called and called in pair.

PiperOrigin-RevId: 561424205