Take screenshots by always passing `afterScreenUpdates:YES`.
Motivation: a lot of tests use `[GREYScreenshotter snapshotElement]` method to take screenshots which currently uses `afterScreenUpdates:NO`. This means that changes done in the current run loop iteration won't be visible in the screenshot. This can cause flakiness and it's better to always use `afterScreenUpdates:YES` to get the most accurate current UI state.
PiperOrigin-RevId: 570316525