Fix the deviation caused by grey_scrollInDirection.
ScrollView's contentOffset cannot be any decimal number. Instead, it's the value that can always be divided by (1 / screen scale).

For example, the latest device has screen scale 3, so the content offset is always one of the x.3333, x.6666 or x.0000. If contentOffset is assigned with a value, the value will then be approximated to the nearest valid number.

However, this will result in EarlGrey scroll action producing deviation on scroll offset result, even if the scroll amount is an integer. For example, when scroll amount is large, EarlGrey will perform scroll gesture multiple times. For each time, it scrolls from the center of the view to the border of the view, which can be a decimal number. This decimal scroll amount will be approximated before applying to contentOffset, but the remainingScrollAmount won't.

To fix the deviation, EarlGrey should always adjust its maximum feasible scroll amount to an integer, which is always a safe value without being approximated.

PiperOrigin-RevId: 595525667