Raise an exception in system alert methods when no error is provided.
It is explicitly stated in the comments/docs of these methods that the failure will be reported, but it was not implemented correctly and the failure has been silently dropped before.

For callers that wish to keep the existing behavior (handle system alerts only if present and do nothing otherwise), they should provide an NSError and not use it afterwards.

Also, simplify handling of the alert and update the error returned -- if a dialog cannot be accepted because no dialog is shown, previously it states that the Accept button is not hittable; now it will state that no system alert is found. Tests are also updated to reflect this change.

PiperOrigin-RevId: 772966694