Introduce rotateLayoutToOrientation:error:
The problem with rotateDeviceToOrientation:error: is that actually
only the layout is rotated, not the device. The method can’t fully
simulate device rotation because it can not consider rotating to face
up or down. By restricting the rotation to layout, it ensures that all
possible non-unknown entries are valid entries.