Change the QoS of the app proxy queue.
The queue is used by the test to wait for requests on the main thread (e.g. from GREYExecuteSyncBlockInBackgroundQueue and a couple other places) to make it appear as if interactions/assertions are happening in a synchronous fashion, so also marking the queue to have QoS user initiated to get rid of priority inversion warning.

PiperOrigin-RevId: 798241134